+++ src/groovy-lint-rules.js
@@ -93,6 +93,8 @@ const rulesFixPriorityOrder = [
     "SpaceBeforeOpeningBrace",
     "SpaceAfterOpeningBrace",
     "SpaceAfterCatch",
+    "SpaceAfterSwitch",
+    "SpaceAfterWhile",
     "SpaceAroundOperator",
     "SpaceAfterComma",
     "SpaceAfterFor",