+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## [3.2.2] 2020-03-31
+
+- New option **returnrules** if you want to return rules descriptions and documentation url in results
+- Use npm ci instead of npm install in CircleCI build
+
 ## [3.2.1] 2020-03-29
 
 - Return rules descriptions in results
-- New option nolintafter: do not lint again a format or a fix, as the client prefers to request it
+- New option **nolintafter**: do not lint again a format or a fix, as the client prefers to request it
 - Fixes
   - [Issue #13](https://github.com/nvuillam/npm-groovy-lint/issues/13): False positive error ClassNameSameAsFileName
   - Sometimes returning wrong .groovylint.json config file